feat: HC-05 initialization protoype
The HC-05 must be configured in what is essentially an admin mode, which can be entered by setting the enable / key pin high, then powering the whole module, configuring, setting the enable pin low, and then power cycling the module. It will then boot into whatever mode (master or slave, baud, etc.) it was configured as. GPIO pins and a pnp BJT will be used to control this configuration sequence purely through software